+++ src/mbgl/util/default_styles.cpp
@@ -4,12 +4,12 @@ namespace mbgl {
 namespace util {
 namespace default_styles {
 
-const DefaultStyle streets   = { "mapbox://styles/mapbox/streets-v8",          "Streets" };
-const DefaultStyle emerald   = { "mapbox://styles/mapbox/emerald-v8",          "Emerald" };
-const DefaultStyle light     = { "mapbox://styles/mapbox/light-v8",            "Light" };
-const DefaultStyle dark      = { "mapbox://styles/mapbox/dark-v8",             "Dark" };
-const DefaultStyle satellite = { "mapbox://styles/mapbox/satellite-v8",        "Satellite" };
-const DefaultStyle hybrid    = { "mapbox://styles/mapbox/satellite-hybrid-v8", "Hybrid" };
+const DefaultStyle streets   = { "mapbox://styles/mapbox/streets-v9",          "Streets" };
+const DefaultStyle outdoors  = { "mapbox://styles/mapbox/outdoors-v9",         "Outdoors" };
+const DefaultStyle light     = { "mapbox://styles/mapbox/light-v9",            "Light" };
+const DefaultStyle dark      = { "mapbox://styles/mapbox/dark-v9",             "Dark" };
+const DefaultStyle satellite = { "mapbox://styles/mapbox/satellite-v9",        "Satellite" };
+const DefaultStyle hybrid    = { "mapbox://styles/mapbox/satellite-hybrid-v9", "Hybrid" };
 
 } // namespace default_styles
 } // end namespace util